+++ src/types/skillbar.ts
@@ -1,0 +1,8 @@
+export interface ProgressOptions {
+    total: number;
+    color: string;
+    width: number;
+    height: number;
+}
+
+export interface CircularSkillBarOptions { elements: string; }